Refactor common CLI code out of cmd_chat_api.go
